Move `./data-list` > `./dl`
