+++ lib/handlers/index.js
@@ -8,7 +8,7 @@ var br = require('./break')
 var cell = require('./table-cell')
 var code = require('./code')
 var comment = require('./comment')
-var dataList = require('./data-list')
+var dl = require('./dl')
 var del = require('./delete')
 var emphasis = require('./emphasis')
 var heading = require('./heading')
@@ -137,7 +137,7 @@ exports.blockquote = blockquote
 exports.br = br
 exports.code = inlineCode
 exports.dir = list
-exports.dl = dataList
+exports.dl = dl
 exports.dt = listItem
 exports.dd = listItem
 exports.del = del
